+++ spec/cpp_library_spec.rb
@@ -40,6 +40,7 @@ RSpec.describe ArduinoCI::CppLibrary do
         "TestSomething/test/good-defines.cpp",
         "TestSomething/test/good-wcharacter.cpp",
         "TestSomething/test/good-wstring.cpp",
+        "TestSomething/test/good-stream.cpp",
         "TestSomething/test/bad-null.cpp",
       ]
       relative_paths = cpp_library.test_files.map { |f| f.split("SampleProjects/", 2)[1] }